Attach public keys to Chameleon accounts
This allows the Ubuntu base image to pull in the SSH public key data added to the vendordata service by
https://github.com/ChameleonCloud/chameleon-vendordata/pull/3. It then appends it to the `authorized_keys`
file of the `cc` user, thus allowing them to immediately `ssh` into any Ubuntu image without having to
manually add their keys.

This patch adds a new post-install script called `50-ssh-copy-public-keys`,
and registers the script to `cloud-cfg.yaml` as an argument to `runcmd`.